pyproject.toml
[tool.poetry]
name = "rss-mounter"
version = "0.1.0"
description = "Mount RSS feed with FUSE."
authors = ["fealone"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
feedparser = "^6.0.2"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
